# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Componode API
  version: 1.0.0
  description: |
    Componode is an open-source, self-hosted Digital Product Asset Management
    (DPAM) tool. This document describes the REST API exposed by the backend
    (`packages/backend`).

    ## Authentication
    Most endpoints require an authenticated session. Sessions are established
    via `POST /api/v1/auth/login` (or OIDC), which sets an `HttpOnly` session
    cookie (`componode_session`). Mutating requests must also send the CSRF
    token returned at login via the `X-CSRF-Token` header (double-submit
    pattern, cookie `componode_csrf`).

    Public endpoints: `POST /auth/login`, `POST /auth/register`,
    `POST /auth/password/reset/confirm`, the OIDC endpoints, and `GET /metrics`
    (no `/api/v1` prefix; network-restricted in production).

    ## Errors
    Error responses use a consistent envelope:
    `{ "code": "<ERROR_CODE>", "message": "...", "details": <optional> }`.
    Error codes are a controlled enum defined in `packages/core`
    (`ERROR_CODES`).

    ## Conventions
    - All routes are prefixed with `/api/v1` except `GET /metrics`.
    - `GET`/`HEAD` routes are side-effect-free on domain state.
    - List queries exclude `RETIRED` records and `GONE` instances unless
      explicitly requested.
    - Entity IDs are UUID v7; human-readable references use `slug`.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: /api/v1
...
paths:
...
  /users:
    get:
      tags: [Users]
      summary: List users
      x-permission: "user:list"
      description: Requires the `user:list` role (Admin).
      parameters:
        - in: query
          name: role
          schema: { $ref: '#/components/schemas/Role' }
        - in: query
          name: isActive
          schema: { type: string, enum: ['true', 'false'] }
        - in: query
          name: search
          schema: { type: string }
      responses:
        '200':
          description: User list
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items: { $ref: '#/components/schemas/Person' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Users]
      summary: Create a user
      x-permission: "user:create"
      description: Requires the `user:create` role (Admin).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password, role]
              properties:
                username:
                  type: string
                  minLength: 3
                  maxLength: 50
                  pattern: '^[a-z0-9_-]+$'
                password: { type: string, minLength: 8, maxLength: 128 }
                role: { $ref: '#/components/schemas/Role' }
                displayName: { type: string, maxLength: 100 }
                email: { type: string, format: email }
                teamId: { type: string, format: uuid }
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: '#/components/schemas/Person' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: Username already taken (`AUTH_USERNAME_TAKEN`)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /users/me:
    get:
      tags: [Users]
      summary: Get the current user's profile
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: '#/components/schemas/Person' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /users/{id}:
    get:
      tags: [Users]
      summary: Get a user by ID
      x-permission: "user:list"
      description: Requires the `user:list` role (Admin).
      parameters:
        - { $ref: '#/components/parameters/Id' }
      responses:
        '200':
          description: User
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: '#/components/schemas/Person' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Users]
      summary: Update a user (role, display name, email, team, active flag)
      x-permission: "user:update"
      description: Requires the `user:update` role (Admin).
      parameters:
        - { $ref: '#/components/parameters/Id' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role: { $ref: '#/components/schemas/Role' }
                displayName: { type: string, maxLength: 100 }
                email: { type: string, format: email }
                teamId: { type: string, format: uuid, nullable: true }
                isActive: { type: boolean }
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: '#/components/schemas/Person' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /settings:
    get:
      tags: [Settings]
      summary: Read application settings
      x-permission: "settings:update"
      description: Requires the `settings:update` role (Admin).
      responses:
        '200':
          description: Settings
          content:
            application/json:
              schema:
                type: object
                properties:
                  settings: { $ref: '#/components/schemas/AppSettings' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    patch:
      tags: [Settings]
      summary: Update application settings
      x-permission: "settings:update"
      description: Requires the `settings:update` role (Admin).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                allowSelfRegistration: { type: boolean }
                sessionIdleTimeoutMs: { type: integer, minimum: 60000 }
                sessionAbsoluteTimeoutMs: { type: integer, minimum: 60000 }
                defaultUserRole: { $ref: '#/components/schemas/Role' }
      responses:
        '200':
          description: Updated settings
          content:
            application/json:
              schema:
                type: object
                properties:
                  settings: { $ref: '#/components/schemas/AppSettings' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /settings/oidc:
    get:
      tags: [Settings]
      summary: Read OIDC configuration
      x-permission: "oidc:configure"
      description: Requires the `oidc:configure` role (Admin).
      responses:
        '200':
          description: OIDC config
          content:
            application/json:
              schema:
                type: object
                properties:
                  oidcConfig: { $ref: '#/components/schemas/OidcConfig' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    put:
      tags: [Settings]
      summary: Update OIDC configuration (tests issuer discovery)
      x-permission: "oidc:configure"
      description: Requires the `oidc:configure` role (Admin).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled]
              properties:
                enabled: { type: boolean }
                issuer: { type: string, format: uri, nullable: true }
                clientId: { type: string, nullable: true }
                clientSecretRef: { type: string, nullable: true }
                roleClaimPath: { type: string, nullable: true }
                claimValueField: { type: string, nullable: true }
                roleMapping:
                  type: object
                  nullable: true
                  additionalProperties: { $ref: '#/components/schemas/Role' }
      responses:
        '200':
          description: Updated OIDC config
          content:
            application/json:
              schema:
                type: object
                properties:
                  oidcConfig: { $ref: '#/components/schemas/OidcConfig' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

  # ---------------------------------------------------------------- Health
...
  /component-groups:
    get:
      tags: [ComponentGroups]
      summary: List component groups
      parameters:
        - in: query
          name: lifecycle
          schema: { $ref: '#/components/schemas/Lifecycle' }
      responses:
        '200':
          description: Component groups
          content:
            application/json:
              schema:
                type: object
                properties:
                  groups:
                    type: array
                    items: { $ref: '#/components/schemas/ComponentGroup' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [ComponentGroups]
      summary: Create a component group
      x-permission: "componentGroup:create"
      description: Requires the `componentGroup:create` role (Editor+).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, slug]
              properties:
                name: { type: string, minLength: 1, maxLength: 100 }
                slug: { type: string, minLength: 1, maxLength: 100 }
                description: { type: string, maxLength: 500 }
                teamOwnerId: { type: string, format: uuid }
      responses:
        '201':
          description: Group created
          content:
            application/json:
              schema:
                type: object
                properties:
                  group: { $ref: '#/components/schemas/ComponentGroup' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: Slug already in use (`SLUG_CONFLICT`)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /component-groups/{id}:
    get:
      tags: [ComponentGroups]
      summary: Get a component group by ID
      parameters:
        - { $ref: '#/components/parameters/Id' }
      responses:
        '200':
          description: Component group
          content:
            application/json:
              schema:
                type: object
                properties:
                  group: { $ref: '#/components/schemas/ComponentGroup' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [ComponentGroups]
      summary: Update a component group
      x-permission: "componentGroup:update"
      description: Requires the `componentGroup:update` role (Editor+).
      parameters:
        - { $ref: '#/components/parameters/Id' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, minLength: 1, maxLength: 100 }
                slug: { type: string, minLength: 1, maxLength: 100 }
                description: { type: string, maxLength: 500 }
                teamOwnerId: { type: string, format: uuid }
                lifecycle: { $ref: '#/components/schemas/Lifecycle' }
      responses:
        '200':
          description: Updated group
          content:
            application/json:
              schema:
                type: object
                properties:
                  group: { $ref: '#/components/schemas/ComponentGroup' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Slug already in use (`SLUG_CONFLICT`)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    delete:
      tags: [ComponentGroups]
      summary: Delete a component group
      x-permission: "componentGroup:delete"
      description: Requires the `componentGroup:delete` role (Editor+).
      parameters:
        - { $ref: '#/components/parameters/Id' }
      responses:
        '204': { description: Group deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # -------------------------------------------------------------- Importers
...
components:
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
      description: Entity ID (UUID v7)
  responses:
    ValidationFailed:
      description: Input validation failed (`VALIDATION_FAILED`)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Not authenticated (`AUTH_NO_SESSION`)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Insufficient role (`AUTH_FORBIDDEN`)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found (`NOT_FOUND`)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - AUTH_INVALID_CREDENTIALS
            - AUTH_RATE_LIMITED
            - AUTH_NO_SESSION
            - AUTH_FORBIDDEN
            - AUTH_USERNAME_TAKEN
            - AUTH_RESET_TOKEN_INVALID
            - AUTH_RESET_TOKEN_EXPIRED
            - AUTH_RESET_TOKEN_USED
            - OIDC_NOT_CONFIGURED
            - OIDC_INVALID_STATE
            - OIDC_INVALID_CODE
            - OIDC_TOKEN_VERIFICATION_FAILED
            - OIDC_DISCOVERY_FAILED
            - CSRF_TOKEN_MISMATCH
            - VALIDATION_FAILED
            - NOT_FOUND
            - CONFLICT
            - CYCLE_DETECTED
            - INVALID_EDGE_TYPE
            - REFERENCED
            - TYPE_CHANGE_BLOCKED
            - SLUG_TAKEN
            - SLUG_CONFLICT
            - CONFIG_NOT_FOUND
            - RUN_IN_PROGRESS
            - RUN_NOT_ACTIVE
            - INTERNAL_ERROR
        message: { type: string }
        details: { description: Optional details (e.g. Zod issues) }
...
    Role:
      type: string
      enum: [ADMIN, EDITOR, VIEWER]
    Lifecycle:
      type: string
      enum: [ACTIVE, RETIRED]
...
    Person:
      type: object
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        oidcSubject: { type: string, nullable: true }
        role: { $ref: '#/components/schemas/Role' }
        displayName: { type: string, nullable: true }
        email: { type: string, format: email, nullable: true }
        teamId: { type: string, format: uuid, nullable: true }
        slug: { type: string }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    AppSettings:
      type: object
      properties:
        allowSelfRegistration: { type: boolean }
        sessionIdleTimeoutMs: { type: integer }
        sessionAbsoluteTimeoutMs: { type: integer }
        defaultUserRole: { $ref: '#/components/schemas/Role' }
    OidcConfig:
      type: object
      properties:
        enabled: { type: boolean }
        issuer: { type: string, nullable: true }
        clientId: { type: string, nullable: true }
        clientSecretRef: { type: string, nullable: true }
        roleClaimPath: { type: string, nullable: true }
        claimValueField: { type: string, nullable: true }
        roleMapping:
          type: object
          nullable: true
          additionalProperties: { type: string }
        updatedAt: { type: string, format: date-time }
...
    ComponentGroup:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        description: { type: string, nullable: true }
        lifecycle: { $ref: '#/components/schemas/Lifecycle' }
        teamOwnerId: { type: string, format: uuid, nullable: true }
        createdBy: { type: string, format: uuid, nullable: true }
        updatedBy: { type: string, format: uuid, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: componode_session
      description: Session cookie set by login/register/OIDC callback
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Required on mutating requests (double-submit CSRF)
security:
  - sessionCookie: []